requirements.txt
asgiref==3.9.1
Django==5.2.4
et_xmlfile==2.0.0
lxml==6.0.0
openpyxl==3.1.5
python-docx==1.2.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
